Tidy up/rearrange/generalize kabuki_extras.
Closing sinks isn't as complicated as I thought it was, and it's a lot
easier to do it by letting the actor's inbox drop than having a
separate oneshot channel for it.

The default TcpStream behavior doesn't actually close the socket when
its sink is closed, though, so it had to be wrapped to add that
behavior.