Initial pass at starting to make an RPC interface.
The problem is that the Session got put in the wrong place in the
abstraction, so I'm committing this WIP to save my work before
rearranging the Session.

As soon as I realized I was adding control messages to the codec
instead of just encoding/decoding, it seemed that it wasn't really a
codec anymore.