Port over to using failure::Error for errors.
In a few cases, this removes a bit of information which was previously
present, but I'll revisit that later. Notably, SendError from actors
no longer provides the message which failed to send, but I have some
ideas for how to retrieve it again, with some sort of thread-local
token. Not sure what that will look like.

Other cases now are _better_ because they return more interesting or
detailed context. Context should be attached in more places, because
otherwise you lose chaining information, but that's again a work in
progress.